Introduce `Param` class.
This fixes a lot of little annoyances I had over the past year or so. It
also changes the interface of all layers with a bias as well as the
`parameters` method. It cleans up the internal code considerable
(e.g. `may_decay` disappears) and will allow to re-initialize an
existing network, which is useful when memory is tight.