Big change in criteria:
Individual criteria now return per-sample costs. The superclass
`Criterion` can weight or mask these in a generic way, and (in the
future) store the individual costs.

This allows any criterion to profit from per-sample weighting and/or
masking and in the future will allow things such as hard negative mining.